+++ src/StkTable/types/index.ts
@@ -164,8 +164,8 @@ export type SortConfig<T extends Record<string, any>> = {
         silent?: boolean;
     };
     /**
-     * string排序是否使用 String.prototype.localCompare
-     * 默认true ($*$应该false)
+     * string sort if use `String.prototype.localCompare`
+     * default: false
      */
     stringLocaleCompare?: boolean;
 };
@@ -187,7 +187,7 @@ export type HighlightConfig = {
  * Configuration options for the sequence column.
  */
 export type SeqConfig = {
-    /** 序号列起始下标 用于适配分页 */
+    /** The initial subscript of the sequence number column is used to adapt the paging. */
     startIndex?: number;
 };
 
@@ -220,3 +220,8 @@ export type HeaderDragConfig<DT extends Record<string, any> = any> =
           /** 禁用拖动的列 */
           disabled?: (col: StkTableColumn<DT>) => boolean;
       };
+
+export type AutoRowHeightConfig<DT> = {
+    /** Estimated row height */
+    expectedHeight?: number | ((row: DT, index: number) => number);
+};